refactor: Update API calls to use axiosInstance for consistency
